Refactored and improved the auth initialization
Creating an observer before every route visit will result in multiple triggers on auth state change. These changes will fix it. Also moved that logic to the service so consumers can also be sure that auth is initialized.

Removed some prototypes and moved them to the service. This will reduce prototype pollution, and confusion.

Removed setting current user in register and login functions, because it will be already set in onAuthStateChanged.